[Bleeding] Don't allow null display names. Fixes BUKKIT-5254
Currently, plugins can set a player's display name to null, which could cause
issues for other calls to getDisplayName that aren't expecting a null value.
This changes setDisplayName to follow the same logic as setPlayerListName,
which sets the player's name back to their unmodified "vanilla" name if it
receives a null value as a parameter.