[Bleeding] Call BlockFadeEvent with proper type in the nether. Fixes BUKKIT-5373
Previously, whenever BlockFadeEvent was called for ice melting, it returned
the new BlockState with a type of Material.STATIONARY_WATER. However, in the
Nether, ice melting does not form water, but is simply replaced by air
instead. This changes the event to use the proper BlockState based on whether
the block is located in the Nether or not.